Fix document renderer architecture gates
